+++ src/jvmMain/java/instructions/direct/JMPDirect.java
@@ -7,7 +7,7 @@ import main.Registers;
 
 public class JMPDirect extends Instruction {
     public JMPDirect() {
-        super("JMP.D", (short) 0xEB, 3);
+        super("JMP.D", (byte) 0xEB, 3);
     }
     public void execute(Registers registers, Memory memory, Short op) {
         short value = memory.getCell(op);
