+++ src/jvmMain/java/instructions/imediate/MOVImediate.java
@@ -8,9 +8,7 @@ import main.Registers;
 
 public class MOVImediate extends Instruction {
 
-    public MOVImediate() {
-        super("MOV.I", (short) 0xA2, 3);
-    }
+    public MOVImediate() { super("MOV.I", (byte) 0xA2, 3); }
     public void execute(Registers registers, Memory memory, Short op) {
         // Memory will not be used in this instruction
         // op is a number
