+++ src/jvmMain/java/instructions/register/POPFRegister.java
@@ -8,7 +8,7 @@ import main.Registers;
 public class POPFRegister extends Instruction {
 
     public POPFRegister() {
-        super("POPF.R", (short) 0x9D, 2);
+        super("POPF.R", (byte) 0x9D, 2);
     }
     public void execute(Registers registers, Memory memory, Short op) {
         Register regDestination = registers.getRegisterByName("SR");
