+++ src/jvmMain/java/instructions/register/POPRegister.java
@@ -8,7 +8,7 @@ import main.Registers;
 public class POPRegister extends Instruction {
 
     public POPRegister() {
-        super("POP.R", (short) 0x58, 2);
+        super("POP.R", (byte) 0x58, 2);
     }
     public void execute(Registers registers, Memory memory, Short op) {
         Register regSource = registers.getRegisterByName("SP");
