+++ src/jvmMain/java/instructions/register/RETRegister.java
@@ -8,7 +8,7 @@ import main.Registers;
 public class RETRegister extends Instruction {
 
     public RETRegister() {
-        super("RET.R", (short) 0xEF, 2);
+        super("RET.R", (byte) 0xEF, 2);
     }
     public void execute(Registers registers, Memory memory, Short op) {
         Register regDestination = registers.getRegisterByName("IP");
